refactor: RandomNumberGenerator 리팩토링 - 움직임 조건 숫자 상수로 명명 - 테스트 코드 위치 이동
