+++ src/main/java/racingcar/domain/RandomNumberGenerator.java
@@ -6,12 +6,13 @@ public class RandomNumberGenerator {
 
     private static final Random RANDOM = new Random();
     private static final int RANDOM_RULE_MAX_NUMBER = 10;
+    public static final int MOVABLE_NUMBER_CONDITION = 4;
 
     public int generate() {
         return RANDOM.nextInt(RANDOM_RULE_MAX_NUMBER);
     }
 
     public boolean isMovable() {
-        return generate() >= 4;
+        return generate() >= MOVABLE_NUMBER_CONDITION;
     }
 }
